Refactor flip recommendation printing

+++ include/Flips.hpp
@@ -46,5 +46,5 @@ namespace flips
 	void filter_count(const db& db, const u32 flip_count);
 
 	/* Flip recommendations */
-	bool flip_recommendations(const db& db, const i64 profit_threshold, const i32 recommendation_count, const i32 random_flip_count, const bool ge_inspector_format);
+	bool flip_recommendations(const db& db, const i64 profit_threshold, const i32 recommendation_count, const size_t random_flip_count, const bool ge_inspector_format);
 }
